% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_validation.R
\name{check_columns_complete}
\alias{check_columns_complete}
\title{Check Completeness of Columns}
\usage{
check_columns_complete(data, vocabulary)
}
\arguments{
\item{data}{The data frame to be checked.}

\item{vocabulary}{The vocabulary containing the index of columns to be checked.}
}
\value{
A boolean indicating if all specified columns are complete (TRUE) or if there are incomplete columns (FALSE).
If there are incomplete columns, a message listing the columns with missing values is returned.
}
\description{
This function checks the completeness of specified columns in a data frame.
It identifies any columns that contain missing values (NA).
}
\examples{
data <- data.frame(
  ID = c(1, 2, NA, 4),
  Name = c("John", "Jane", "Tom", "Mary"),
  Age = c(25, 30, NA, 40)
)
vocabulary <- list(
  index = list(
    checkCols = c("ID", "Age")
  )
)
result <- check_columns_complete(data, vocabulary)
if (result) {
  # All columns are complete
  print("All columns are complete.")
} else {
  # Incomplete columns exist
  print(result)
}

}
