% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_validation.R
\name{check_rows_complete}
\alias{check_rows_complete}
\title{Check Completeness of Rows}
\usage{
check_rows_complete(data, vocabulary)
}
\arguments{
\item{data}{The input data frame.}

\item{vocabulary}{The vocabulary containing the columns to check.}
}
\value{
TRUE if all rows are complete, or an error message indicating the rows and columns with missing values.
}
\description{
This function checks the completeness of rows in a data frame based on a given vocabulary.
It verifies if the specified columns in each row have missing values (NA).
If any rows are found to be incomplete, it returns an error message indicating the rows and columns with missing values.
If all rows are complete, it returns TRUE.
}
\examples{
data <- data.frame(
  A = c(1, NA, 3),
  B = c("x", "y", NA),
  C = c("a", "b", "c")
)
vocabulary <- list(
  index = list(
    checkCols = c("A", "B"),
    checkRows = c(1, 2, 3)
  )
)
result <- check_rows_complete(data, vocabulary)
if (isTRUE(result)) {
  # All rows are complete
  print("All rows are complete.")
} else {
  # Incomplete rows
  print(result)
}

}
