% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_validation.R
\name{check_vocabulary_data}
\alias{check_vocabulary_data}
\title{Check Vocabulary Data}
\usage{
check_vocabulary_data(data, vocabulary)
}
\arguments{
\item{data}{The data frame to be checked.}

\item{vocabulary}{The vocabulary containing the allowed values for each column.}
}
\value{
A boolean indicating if all values in the specified columns are allowed (TRUE) or if there are disallowed values (FALSE).
If there are disallowed values, an error message is returned indicating the column and the allowed values.
}
\description{
This function checks if the values in specified columns of a data frame are allowed based on a vocabulary.
It compares the values in each column with the allowed values specified in the vocabulary.
If any values are found that are not allowed, an error message is returned.
}
\examples{
data <- data.frame(
  ID = c(1, 2, 3, 4),
  Status = c("Active", "Inactive", "Unknown", "Active")
)
vocabulary <- list(
  vocabulary = list(
    Status = c("Active", "Inactive")
  )
)
result <- check_vocabulary_data(data, vocabulary)
if (result) {
  # All values in specified columns are allowed
  print("All values are allowed.")
} else {
  # Disallowed values exist
  print(result)
}

}
