% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_validation.R
\name{check_vocabulary_metadata}
\alias{check_vocabulary_metadata}
\title{Check Vocabulary Metadata}
\usage{
check_vocabulary_metadata(data, vocabulary)
}
\arguments{
\item{data}{The input data frame.}

\item{vocabulary}{The vocabulary containing the allowed values for each column.}
}
\value{
TRUE if all values are allowed, or an error message indicating the row, column, and allowed values.
}
\description{
This function checks the metadata of a data frame against a given vocabulary.
It verifies if the values in specified columns of each row are allowed according to the vocabulary.
If any value is found to be not allowed, it returns an error message indicating the row, column, and allowed values.
If all values are allowed, it returns TRUE.
}
\examples{
data <- data.frame(
  A = c("x", "y", "z"),
  B = c("a", "b", "c")
)
vocabulary <- list(
  index = list(
    checkCols = c("A", "B")
  ),
  vocabulary = list(
    A = c("x", "y"),
    B = c("a", "c")
  )
)
result <- check_vocabulary_metadata(data, vocabulary)
if (isTRUE(result)) {
  # All values are allowed
  print("All values are allowed.")
} else {
  # Disallowed values
  print(result)
}

}
