% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_data_processing.R
\name{fromshapetowkt}
\alias{fromshapetowkt}
\title{Convert from Shapefile to WKT format}
\usage{
fromshapetowkt(data, shape, id_data_colmn, id_shape_colmn, geom_colmn)
}
\arguments{
\item{data}{The dataset to be merged with the shapefile.}

\item{shape}{The shapefile object.}

\item{id_data_colmn}{The column name in the dataset representing the location ID.}

\item{id_shape_colmn}{The column name in the shapefile representing the location ID.}

\item{geom_colmn}{The column name in the shapefile representing the geometry.}
}
\value{
A spatial data frame (sf) containing the merged data with the WKT column.
}
\description{
This function takes a dataset and a shapefile and performs the following steps:
1. Selects the specified columns from the shapefile.
2. Renames the selected column to "locationID".
3. Merges the dataset and the modified shapefile based on the "locationID" column.
4. Converts the merged data into a spatial data frame (sf) using the specified geometry column.
5. Assigns the Coordinate Reference System (CRS) from the shapefile to the merged sf object.
6. Returns the merged data with the WKT column.
}
\examples{
data <- read.csv("data.csv")
shape <- sf::read_sf("shapefile.shp")
merged_data <- fromshapetowkt(data, shape, "locationID", "locationID", "geometry")

}
