% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_utils.R
\name{generate_report}
\alias{generate_report}
\title{Generate report in modal}
\usage{
generate_report(report_title, report_text, btn_id, btn_name, btn_type)
}
\arguments{
\item{report_title}{The title of the report modal dialog.}

\item{report_text}{The text content of the report.}

\item{btn_id}{The ID of the button.}

\item{btn_name}{The label/name of the button.}

\item{btn_type}{The type of the button ("action" or "download").}
}
\description{
This function generates a report in a modal dialog with a specified title, text, and button.
The button type can be either an action button or a download button, based on the value of `btn_type`.
The generated report is displayed as pre-formatted text in the modal dialog.
}
\examples{
report_title <- "Report"
report_text <- c("Line 1", "Line 2", "Line 3")
btn_id <- "btn1"
btn_name <- "Download"
btn_type <- "download"
generate_report(report_title, report_text, btn_id, btn_name, btn_type)

}
