% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_data_processing.R
\name{read_data_global}
\alias{read_data_global}
\title{Read global data from Excel file}
\usage{
read_data_global(inFile_DATA)
}
\arguments{
\item{inFile_DATA}{The input file containing the global data.}
}
\value{
The global data read from the Excel file.
}
\description{
This function reads global data from an Excel file using the `read_excel_allsheets` function.
}
\examples{
inFile <- file.choose()
data <- read_data_global(inFile)

}
\seealso{
\code{\link{read_excel_allsheets}}
}
