% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_data_processing.R
\name{read_gpkg_global}
\alias{read_gpkg_global}
\title{Read global data from GeoPackage file}
\usage{
read_gpkg_global(inFile_GPKG)
}
\arguments{
\item{inFile_GPKG}{The input file containing the GeoPackage data.}
}
\value{
The global data read from the GeoPackage file as a spatial data frame (sf).
}
\description{
This function reads global data from a GeoPackage file using the `read_sf` function from the 'sf' package.
}
\examples{
inFile <- file.choose()
data <- read_gpkg_global(inFile)

}
