% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_data_processing.R
\name{read_shapefile}
\alias{read_shapefile}
\title{Read shapefile}
\usage{
read_shapefile(shp_path)
}
\arguments{
\item{shp_path}{The path to the shapefile.}
}
\value{
A spatial data frame (sf) representing the contents of the shapefile.
}
\description{
This function reads a shapefile from the specified path using the 'sf' package.
It renames the files in the shapefile directory to ensure consistency, and then reads the shapefile into a spatial data frame.
}
\examples{
shp_path <- list(datapath = "path/to/shapefile", name = "shapefile.shp")
data <- read_shapefile(shp_path)

}
