% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_validation.R
\name{submission_panel_test}
\alias{submission_panel_test}
\title{Submission Panel Test}
\usage{
submission_panel_test(
  revised_by_surname,
  revised_by_lastname,
  revised_by_mail,
  revised_by_institution,
  version,
  shape_format,
  num_shapefiles,
  data_file,
  shape_file,
  id_column,
  geom_column,
  data_agreement_selection,
  data_agreement
)
}
\arguments{
\item{revised_by_surname}{Surname of the reviser.}

\item{revised_by_lastname}{Lastname of the reviser.}

\item{revised_by_mail}{Email of the reviser.}

\item{revised_by_institution}{Name of the institution.}

\item{version}{Version.}

\item{shape_format}{Shapefile format.}

\item{num_shapefiles}{Number of shapefiles.}

\item{data_file}{Data file.}

\item{shape_file}{Spatial file.}

\item{id_column}{ID column from the spatial file.}

\item{geom_column}{Geometry column from the spatial file.}

\item{data_agreement_selection}{Data agreement selection ("Yes" or "No").}

\item{data_agreement}{Data agreement file.}
}
\value{
A character vector containing the missing fields, or TRUE if all required fields are provided.
}
\description{
This function performs a validation test on the submission panel form inputs.
It checks for the presence of required fields and returns a list of missing fields if any are found.
If all required fields are provided, it returns TRUE.
}
\examples{
revised_by_surname <- "Doe"
revised_by_lastname <- "John"
revised_by_mail <- "johndoe@example.com"
revised_by_institution <- "Example Institution"
version <- "1.0"
shape_format <- "wktxy"
num_shapefiles <- 2
data_file <- "data.csv"
shape_file <- "shapefile.zip"
id_column <- "ID"
geom_column <- "Geometry"
data_agreement_selection <- "Yes"
data_agreement <- "agreement.pdf"
result <- submission_panel_test(revised_by_surname, revised_by_lastname, revised_by_mail, revised_by_institution, version, shape_format, num_shapefiles, data_file, shape_file, id_column, geom_column, data_agreement_selection, data_agreement)
if (isTRUE(result)) {
  # All required fields are provided
  print("Submission panel is valid.")
} else {
  # Missing fields
  print(paste("Missing fields:", result))
}

}
