% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_validation.R
\name{validate_match}
\alias{validate_match}
\title{Validate Match between Data and Shape}
\usage{
validate_match(data_uploaded, shape_uploaded, id_column, id_column_shape)
}
\arguments{
\item{data_uploaded}{The data frame containing the uploaded data.}

\item{shape_uploaded}{The data frame containing the uploaded shape file.}

\item{id_column}{The name of the ID column in the data file.}

\item{id_column_shape}{The name of the ID column in the shape file.}
}
\value{
A list with two elements:
  - The first element is a boolean indicating if the match is valid (TRUE) or not (FALSE).
  - The second element is a message explaining the validation result.
}
\description{
This function validates the match between the data and shape files based on the specified ID columns.
It checks if all the unique location IDs in the data file have a corresponding ID in the shape file.
}
\examples{
data <- data.frame(
  locationID = c("A", "B", "C", "D"),
  value = c(1, 2, 3, 4)
)
shape <- data.frame(
  ID = c("A", "C", "D"),
  geometry = c("POINT", "POLYGON", "POLYGON")
)
result <- validate_match(data, shape, "locationID", "ID")
if (result[[1]]) {
  # Match is valid
  print(result[[2]])
} else {
  # Match is not valid
  print(result[[2]])
}

}
